# 5x faster `toString("base64url")`