### What does this PR do?

Blocked on #11085 

Uses SIMDUTF for `toString("base64url")`, cc @lemire . I'm not confident the function produces output that is 100% consistent with Node. We already had copied most of the `"base64url"` tests though

New:
```ts
cpu: Apple M3 Max
runtime: bun 1.1.9 (arm64-darwin)

benchmark                                 time (avg)             (min … max)       p75       p99      p995
---------------------------------------------------------------------------- -----------------------------
Buffer(110000).toString('base64url')    7.24 µs/iter     (4.54 µs … 1.67 ms)   5.75 µs  15.42 µs  33.67 µs
Buffer(16).toString('base64url')       32.89 ns/iter   (29.29 ns … 90.13 ns)  32.87 ns  43.29 ns  46.56 ns
```

Bun v1.1.7:
```ts
cpu: Apple M3 Max
runtime: bun 1.1.7 (arm64-darwin)

benchmark                                 time (avg)             (min … max)       p75       p99      p995
---------------------------------------------------------------------------- -----------------------------
Buffer(110000).toString('base64url')   39.89 µs/iter  (34.71 µs … 801.13 µs)  39.08 µs   54.5 µs  77.13 µs
Buffer(16).toString('base64url')       29.28 ns/iter   (25.94 ns … 86.85 ns)  29.27 ns  37.24 ns  39.98 ns
```

Node $HEAD:
```ts
cpu: Apple M3 Max
runtime: node v23.0.0-head (arm64-darwin)

benchmark                                 time (avg)             (min … max)       p75       p99      p995
---------------------------------------------------------------------------- -----------------------------
Buffer(110000).toString('base64url')   23.02 µs/iter     (16 µs … 223.25 µs)  18.92 µs  90.08 µs   96.5 µs
Buffer(16).toString('base64url')       55.88 ns/iter    (53.6 ns … 79.34 ns)   57.1 ns  69.57 ns  72.36 ns
```

Deno:
```ts
cpu: unknown
runtime: deno 1.43.3 (aarch64-apple-darwin)

benchmark                                 time (avg)             (min … max)       p75       p99      p995
---------------------------------------------------------------------------- -----------------------------
Buffer(110000).toString('base64url')   43.25 µs/iter    (34.83 µs … 3.95 ms)  37.79 µs  97.88 µs 102.83 µs
Buffer(16).toString('base64url')       221.1 ns/iter (211.16 ns … 247.57 ns) 223.85 ns 237.72 ns 243.17 ns
```

### How did you verify your code works?

Ran buffer & string decoder tests